code extracted to helper class to avoid duplication
