+++ src/main/java/uk/gov/hmcts/sptribs/caseworker/event/CaseworkerEditRecordListing.java
@@ -14,6 +14,7 @@ import uk.gov.hmcts.reform.ccd.client.model.SubmittedCallbackResponse;
 import uk.gov.hmcts.sptribs.caseworker.event.page.RecordNotifyParties;
 import uk.gov.hmcts.sptribs.caseworker.event.page.SelectTemplate;
 import uk.gov.hmcts.sptribs.caseworker.event.page.UploadHearingNotice;
+import uk.gov.hmcts.sptribs.caseworker.helper.RecordListingHelper;
 import uk.gov.hmcts.sptribs.caseworker.model.RecordListing;
 import uk.gov.hmcts.sptribs.ciccase.model.CaseData;
 import uk.gov.hmcts.sptribs.ciccase.model.CicCase;
@@ -30,6 +31,7 @@ import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+
 
 import static uk.gov.hmcts.sptribs.caseworker.util.EventConstants.HYPHEN;
 import static uk.gov.hmcts.sptribs.ciccase.model.State.AwaitingHearing;
@@ -51,6 +53,8 @@ public class CaseworkerEditRecordListing implements CCDConfig<CaseData, State, U
 
     @Autowired
     private LocationService locationService;
+    @Autowired
+    private RecordListingHelper recordListingHelper;
 
     @Override
     public void configure(ConfigBuilder<CaseData, State, UserRole> configBuilder) {
@@ -67,7 +71,7 @@ public class CaseworkerEditRecordListing implements CCDConfig<CaseData, State, U
             .grant(CREATE_READ_UPDATE_DELETE, COURT_ADMIN_CIC, SUPER_USER)
             .grantHistoryOnly(SOLICITOR));
 
-        addHearingTypeAndFormat(pageBuilder);
+        RecordListingHelper.addHearingTypeAndFormat(pageBuilder);
         addRegionInfo(pageBuilder);
         hearingVenues.addTo(pageBuilder);
         addRemoteHearingInfo(pageBuilder);
@@ -80,13 +84,7 @@ public class CaseworkerEditRecordListing implements CCDConfig<CaseData, State, U
 
     public AboutToStartOrSubmitResponse<CaseData, State> aboutToStart(CaseDetails<CaseData, State> details) {
         var caseData = details.getData();
-        DynamicList regionList = locationService.getAllRegions();
-        caseData.getRecordListing().setRegionList(regionList);
-
-        String regionMessage = regionList == null || regionList.getListItems().isEmpty()
-            ? "Unable to retrieve Region data"
-            : null;
-        caseData.getRecordListing().setRegionsMessage(regionMessage);
+        RecordListingHelper.caseDataList(caseData, locationService);
 
         return AboutToStartOrSubmitResponse.<CaseData, State>builder()
             .data(caseData)
@@ -94,34 +92,30 @@ public class CaseworkerEditRecordListing implements CCDConfig<CaseData, State, U
             .build();
     }
 
+//    private void caseDataList(CaseData caseData) {
+//        DynamicList regionList = locationService.getAllRegions();
+//        caseData.getRecordListing().setRegionList(regionList);
+//
+//        String regionMessage = regionList == null || regionList.getListItems().isEmpty()
+//            ? "Unable to retrieve Region data"
+//            : null;
+//        caseData.getRecordListing().setRegionsMessage(regionMessage);
+//    }
+
     @SneakyThrows
     public AboutToStartOrSubmitResponse<CaseData, State> aboutToSubmit(CaseDetails<CaseData, State> details,
                                                                        CaseDetails<CaseData, State> beforeDetails) {
+        var caseData = details.getData();
         log.info("Caseworker record listing callback invoked for Case Id: {}", details.getId());
 
-        var caseData = details.getData();
-        final List<String> errors = new ArrayList<>();
-
-        if (checkNullCondition(details.getData().getCicCase())) {
-            errors.add("One party must be selected.");
-        }
-        Set<NotificationParties> partiesSet = new HashSet<>();
-        if (!CollectionUtils.isEmpty(caseData.getCicCase().getRecordNotifyPartySubject())) {
-            partiesSet.add(NotificationParties.SUBJECT);
-        }
-        if (!CollectionUtils.isEmpty(caseData.getCicCase().getRecordNotifyPartyRepresentative())) {
-            partiesSet.add(NotificationParties.REPRESENTATIVE);
-        }
-        if (!CollectionUtils.isEmpty(caseData.getCicCase().getRecordNotifyPartyRespondent())) {
-            partiesSet.add(NotificationParties.RESPONDENT);
-        }
-        caseData.getRecordListing().setNotificationParties(partiesSet);
+        final List<String> errors = RecordListingHelper.partySelector(details, caseData);
         return AboutToStartOrSubmitResponse.<CaseData, State>builder()
             .data(caseData)
             .state(AwaitingHearing)
             .errors(errors)
             .build();
     }
+
 
     public SubmittedCallbackResponse submitted(CaseDetails<CaseData, State> details,
                                                CaseDetails<CaseData, State> beforeDetails) {
@@ -133,40 +127,14 @@ public class CaseworkerEditRecordListing implements CCDConfig<CaseData, State, U
     public AboutToStartOrSubmitResponse<CaseData, State> midEvent(CaseDetails<CaseData, State> details,
                                                                   CaseDetails<CaseData, State> detailsBefore) {
         final CaseData caseData = details.getData();
-        String selectedRegion = caseData.getRecordListing().getSelectedRegionVal();
-        String regionId = getRegionId(selectedRegion);
-
-        if (null != regionId) {
-            DynamicList hearingVenueList = locationService.getHearingVenuesByRegion(regionId);
-            caseData.getRecordListing().setHearingVenues(hearingVenueList);
-
-            String hearingVenueMessage = hearingVenueList == null || hearingVenueList.getListItems().isEmpty()
-                ? "Unable to retrieve Hearing Venues data"
-                : null;
-            caseData.getRecordListing().setHearingVenuesMessage(hearingVenueMessage);
-
-        }
-
+        recordListingHelper.populatedRegionData(caseData,locationService);
         return AboutToStartOrSubmitResponse.<CaseData, State>builder()
             .data(caseData)
             .build();
     }
 
-    private boolean checkNullCondition(CicCase cicCase) {
-        return null != cicCase
-            && CollectionUtils.isEmpty(cicCase.getRecordNotifyPartySubject())
-            && CollectionUtils.isEmpty(cicCase.getRecordNotifyPartyRepresentative())
-            && CollectionUtils.isEmpty(cicCase.getRecordNotifyPartyRespondent());
-    }
 
-    private void addHearingTypeAndFormat(PageBuilder pageBuilder) {
-        pageBuilder.page("hearingTypeAndFormat")
-            .pageLabel("Hearing type and format")
-            .complex(CaseData::getRecordListing)
-            .mandatory(RecordListing::getHearingType)
-            .mandatory(RecordListing::getHearingFormat)
-            .done();
-    }
+
 
     private void addRegionInfo(PageBuilder pageBuilder) {
         pageBuilder.page("regionInfo", this::midEvent)
@@ -177,12 +145,6 @@ public class CaseworkerEditRecordListing implements CCDConfig<CaseData, State, U
             .done();
     }
 
-    private String getRegionId(String selectedRegion) {
-        String[] values = selectedRegion != null
-            ? Arrays.stream(selectedRegion.split(HYPHEN)).map(String::trim).toArray(String[]::new)
-            : null;
-        return values != null && values.length > 0 ? values[0] : null;
-    }
 
     private void addRemoteHearingInfo(PageBuilder pageBuilder) {
         pageBuilder.page("remoteHearingInformation")
